Guess what I found :)
Negascout - outperforms alpha-beta GIVEN the ordering of child (actions) are good. I presume ours will (or can) be good.

Exit -> Jump (captures) -> Moves -> Pass

For jumps, we can call Paris to see if our jump is a capturing jump, and do .insert(0, piece) to insert it into the front of the jump actions